requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
argcomplete==1.10.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.8.2
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
compressed-rtf==1.0.7
docx2txt==0.9
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.28.7
greenlet==3.2.3
h11==0.16.0
idna==3.10
IMAPClient==2.1.0
inflate64==1.0.3
joblib==1.5.1
kombu==5.5.4
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
multivolumefile==0.2.3
nltk==3.9.1
nodeenv==1.9.1
olefile==0.47
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20191110
pillow==11.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
py7zr==1.0.0
pybcj==1.0.6
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.8.0
PyPDF2==3.0.1
pyppmd==1.2.0
pyright==1.1.403
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-magic==0.4.27
python-pptx==0.6.23
PyYAML==6.0.2
pyzstd==0.17.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
selenium==4.34.2
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
textract==1.6.5
texttable==1.7.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlrd==1.2.0
xlsxwriter==3.2.5
